fix: replace blog post content with rich-text
