+++ package.json
@@ -8,6 +8,7 @@
     "@babel/plugin-syntax-typescript": "^7.12.13",
     "@loadable/component": "^5.13.1",
     "datocms-structured-text-to-html-string": "^1.1.3",
+    "datocms-structured-text-to-plain-text": "^2.0.4",
     "gatsby": "^2.32.4",
     "gatsby-awesome-pagination": "^0.3.6",
     "gatsby-background-image": "^1.1.2",
