+++ src/utils/render-options.js
@@ -1,0 +1,16 @@
+const options = {
+  renderBlock({ record, adapter: { renderNode } }) {
+    return renderNode(
+      "figure",
+      {},
+      renderNode("img", { src: record.image.url, alt: record.image.alt }),
+      renderNode(
+        "figcaption",
+        { style: "text-align: center" },
+        record.image.title
+      )
+    )
+  },
+}
+
+export default options
